♻️  REFACTOR: Move `NodeComments` to separate file
